Moved traits in types to solve name resolution issues
